refactor: moving toast hook to context
