--- src/hooks/useToast.ts
@@ -1,12 +1,0 @@
-import { useContext } from "react";
-
-import { ToastContext, ToastContextType } from "@/contexts/ToastContext";
-
-export const useToast = (): ToastContextType => {
-    const context = useContext(ToastContext);
-    if (!context) {
-      throw new Error('useToast must be used within a ToastProvider');
-    }
-    return context;
-  };
-
